🐛Set focus to input when editing
