+++ src/components/TodoItem.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { useOvermind } from '../overmind'
 import { Todo } from '../overmind/state'
+import TodoItemEdit from './TodoItemEdit'
 
 interface Props {
   todo: Todo
@@ -8,25 +9,6 @@ interface Props {
 
 const TodoItem = ({ todo }: Props) => {
   const { actions, state } = useOvermind()
-  const [title, setTitle] = React.useState(todo.title)
-
-  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(event.target.value)
-  }
-
-  const handleKeyUp = (event: React.KeyboardEvent) => {
-    if (event.keyCode === 27) {
-      actions.stopEditing()
-      return
-    }
-    if (event.keyCode !== 13) {
-      return
-    }
-    event.preventDefault()
-    if (title) {
-      actions.saveEditingTodo(title)
-    }
-  }
 
   const handleClick = () => {
     actions.toggleTodo(todo.id)
@@ -40,12 +22,7 @@ const TodoItem = ({ todo }: Props) => {
   return (
     <li>
       {state.editingTodoId === todo.id ? (
-        <input
-          type="text"
-          defaultValue={todo.title}
-          onChange={handleTitleChange}
-          onKeyUp={handleKeyUp}
-        />
+        <TodoItemEdit todo={todo} />
       ) : (
         <div>
           <input
